Update class for day method and insert methods
todays classes is really just a shortcut to get to classes for a given
day. Created a class method that handles querying for this. todays
classes leverages this method and simply passes current time.

Insert new method now takes an array. If for whatever reason an
exception is thrown during processing, rescue to "next"